rename Mixer to MixerBase to resolve case-insensitivity on mac
